requirements.txt
requests
twilio
flask
websockets
python-dotenv
openai